requirements.txt
Pillow
Django
psycopg2-binary
djangorestframework